Candidate_Service/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
pyjwt
argon2-cffi
sqlalchemy
asyncpg
python-dotenv
icecream
aiosmtplib
orjson